+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.6.0-rc.2]: https://github.com/trussed-dev/ctap-types/compare/0.6.0-rc.1...0.6.0-rc.2
 
 - Add `test-client` feature for additional `Serialize` and `Deserialize` implementations not required by authenticators.
+- Rename `authenticator_config` to `config`.
 
 ## [0.6.0-rc.1] 2026-05-21