+++ CHANGELOG.md
@@ -6,7 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-[Unreleased]: https://github.com/trussed-dev/ctap-types/compare/0.6.0-rc.1...HEAD
+[Unreleased]: https://github.com/trussed-dev/ctap-types/compare/0.6.0-rc.2...HEAD
+
+-
+
+## [0.6.0-rc.2] 2026-05-27
+
+[0.6.0-rc.1]: https://github.com/trussed-dev/ctap-types/compare/0.6.0-rc.2...0.6.0-rc.2
 
 - Rename `authenticator_config` to `config`.
 - Add `platform-serde` feature for additional `Serialize` and `Deserialize` implementations not required by authenticators.